add target flags
