+++ packages/react-devtools-shared/src/hook.js
@@ -11,6 +11,7 @@ import {
   patch as patchConsole,
   registerRenderer as registerRendererWithConsole,
 } from './backend/console';
+import {enableConsolePatching} from 'react-devtools-target-flags';
 
 import type {DevToolsHook} from 'react-devtools-shared/src/backend/types';
 
@@ -172,7 +173,7 @@ export function installHook(target: any): DevToolsHook | null {
     // In that case, we'll patch later (when the frontend attaches).
     //
     // Don't patch in test environments because we don't want to interfere with Jest's own console overrides.
-    if (process.env.NODE_ENV !== 'test' && process.env.TARGET !== 'extension') {
+    if (enableConsolePatching && process.env.NODE_ENV !== 'test') {
       try {
         const appendComponentStack =
           window.__REACT_DEVTOOLS_APPEND_COMPONENT_STACK__ !== false;
