+++ scripts/flow/createFlowConfigs.js
@@ -53,6 +53,7 @@ module.name_mapper='ReactServerStreamConfig$$' -> 'forks/ReactServerStreamConfig
 module.name_mapper='ReactServerFormatConfig$$' -> 'forks/ReactServerFormatConfig.${serverRenderer}'
 module.name_mapper='ReactFlightServerConfig$$' -> 'forks/ReactFlightServerConfig.${serverRenderer}'
 module.name_mapper='ReactFlightClientHostConfig$$' -> 'forks/ReactFlightClientHostConfig.${serverRenderer}'
+module.name_mapper='react-devtools-target-flags' -> 'react-devtools-shared/config/TargetFlags.default'
     `.trim(),
     )
     .replace('%REACT_RENDERER_FLOW_IGNORES%', ignoredPaths.join('\n'));
